Move heatmap component in separate folder
