+++ app/components/pool/graphs/heatmap/heatmap-variables.scss
