+++ app/components/pool/graphs/heatmap/index.ts
@@ -1,0 +1,3 @@
+export * from "./nodes-heatmap.component";
+export * from "./legend";
+export * from "./state-counter";
