+++ app/components/pool/graphs/heatmap/legend/index.ts
@@ -1,0 +1,1 @@
+export * from "./nodes-heatmap-legend.component";
