+++ app/components/pool/graphs/heatmap/legend/nodes-heatmap-legend.html
