+++ app/components/pool/graphs/heatmap/legend/nodes-heatmap-legend.scss
@@ -1,5 +1,5 @@
 @import "app/styles/variables";
-@import "./heatmap-variables";
+@import "../heatmap-variables";
 
 bl-nodes-heatmap-legend {
     width: $legend-width;
