+++ app/components/pool/graphs/heatmap/nodes-heatmap.scss
