+++ app/components/pool/graphs/heatmap/state-counter.ts
