+++ app/components/pool/graphs/index.ts
@@ -1,4 +1,3 @@
 export * from "./pool-graphs.module";
-export * from "./nodes-heatmap.component";
-export * from "./nodes-heatmap-legend.component";
+export * from "./heatmap";
 export * from "./pool-graphs.component";
