+++ app/components/pool/graphs/pool-graphs.component.ts
@@ -11,9 +11,9 @@ import { Job, JobState, Node, NodeState, Pool, Task } from "app/models";
 import { AppInsightsQueryService, JobListParams, JobService, NodeListParams, NodeService } from "app/services";
 import { ListView, PollObservable, PollService } from "app/services/core";
 import { FilterBuilder } from "app/utils/filter-builder";
+import { StateCounter } from "./heatmap";
 import { NodesStateHistoryData, RunningTasksHistoryData } from "./history-data";
 import "./pool-graphs.scss";
-import { StateCounter } from "./state-counter";
 
 enum AvailableGraph {
     Heatmap,
