+++ app/components/pool/graphs/pool-graphs.module.ts
@@ -5,9 +5,8 @@ import { RouterModule } from "@angular/router";
 import { MaterialModule } from "app/core";
 
 import { BaseModule } from "app/components/base";
+import { NodesHeatmapComponent, NodesHeatmapLegendComponent } from "./heatmap";
 import { HistoryGraphComponent } from "./history-graph";
-import { NodesHeatmapLegendComponent } from "./nodes-heatmap-legend.component";
-import { NodesHeatmapComponent } from "./nodes-heatmap.component";
 import {
     CpuUsageGraphComponent, DiskUsageGraphComponent, EnableAppInsightsDocComponent,
     MemoryUsageGraphComponent, NetworkUsageGraphComponent, PerformanceGraphComponent,
