+++ app/components/pool/graphs/pool-graphs.scss
@@ -71,7 +71,7 @@ bl-pool-graphs {
     .full-display-heatmap {
         display: block;
         width: 100%;
-        height: 600px;
+        height: 100%;
     }
 
     .focused-graph {
